+++ witherable/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added `ordNubBy`, `hashNubBy`, `ordNubByOf`, and `hashNubByOf`.
 * Use `alterF` for nub-function implementations
 * Implement `witherM` in `Witherable Vector` instance.
+* Mark modules as Trustworthy
 
 0.4
 -------